+++ Marindeck/Application/AppDelegate.swift
@@ -8,6 +8,7 @@
 import UIKit
 import Keys
 import GiphyUISDK
+import SwiftyStoreKit
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
@@ -30,6 +31,23 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         #if DEBUG
         _ = Test()
         #endif
+        
+        SwiftyStoreKit.completeTransactions(atomically: true) { purchases in
+            for purchase in purchases {
+                switch purchase.transaction.transactionState {
+                case .purchased, .restored:
+                    if purchase.needsFinishTransaction {
+                        // Deliver content from server, then:
+                        SwiftyStoreKit.finishTransaction(purchase.transaction)
+                    }
+                    // Unlock content
+                case .failed, .purchasing, .deferred:
+                    break // do nothing
+                @unknown default:
+                    break
+                }
+            }
+        }
 
         return true
     }
